refactor: extract node extensions to a public class
