+++ Backend/src/Modules/Nodes/Nodes.Application/Entities/Nodes/Extensions/NodeExtensions.cs
@@ -1,0 +1,19 @@
+using Nodes.Application.Entities.Nodes.Dtos;
+
+namespace Nodes.Application.Entities.Nodes.Extensions;
+
+public static class NodeExtensions
+{
+    public static NodeDto ToNodeDto(this Node node)
+    {
+        return new NodeDto(
+            node.Id.ToString(),
+            node.Title,
+            node.Description,
+            node.Timestamp,
+            node.Importance,
+            node.Phase,
+            node.Categories.ToList(),
+            node.Tags.ToList());
+    }
+}
